# products/swagger.yaml
swagger: '2.0'
info:
  description: API Rest de Produto. Versão pré-correção. (Algumas classes de teste foram ocultadas da documentação.)
  version: '1.0'
  title: MS-Product - Programa de Bolsa API Documentation
  termsOfService: urn:tos
  contact:
    name: Joao Pedro Cavalheiro
    email: joao.cavalheiro.pb@compasso.com.br
host: localhost:9999
basePath: /
paths:
  /products:
    get:
      tags:
        - product-controller
      summary: list
      operationId: listUsingGET
      produces:
        - '*/*'
      parameters:
        - name: nameProduct
          in: query
          description: nameProduct
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ProductDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - product-controller
      summary: create
      operationId: createUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: form
          description: form
          required: true
          schema:
            $ref: '#/definitions/ProductForm'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProductDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /products/{id}:
    get:
      tags:
        - product-controller
      summary: findById
      operationId: findByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - product-controller
      summary: update
      operationId: updateUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: form
          description: form
          required: true
          schema:
            $ref: '#/definitions/UptadeProductForm'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProductDto'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - product-controller
      summary: delete
      operationId: deleteUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
definitions:
  ProductDto:
    type: object
    properties:
      description:
        type: string
      id:
        type: integer
        format: int64
      name:
        type: string
      price:
        type: number
        format: double
    title: ProductDto
  ProductForm:
    type: object
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
        format: double
    title: ProductForm
  UptadeProductForm:
    type: object
    properties:
      description:
        type: string
      name:
        type: string
    title: UptadeProductForm